게시글 삭제2

+++ src/main/resources/static/js/boardMain.js
@@ -1,21 +1,27 @@
-const main = {
+const board = {
     init : function () {
-        var _this = this;
-        $('#btn-delete').on('click',function () {
-            _this.delete();
+        $('#btn-board-delete').on('click', () =>{
+            this.deleteById();
         });
     },
-    delete : function () {
-        $.ajax({
-            type: 'DELETE',
-            url : '/board/boardDelete',
-        }).done(function () {
-            alert('모든 글 삭제');
-            window.location.href = '/';
-        }).fail(function (error){
-            alert(JSON.stringify(error));
-        })
-    }
+    deleteById : function () {
+
+        let id = $("#boardID").val();
+
+        if (confirm("정말로 삭제하시겠습니까?")) {
+            $.ajax({
+                type: 'DELETE',
+                url: `/api/v1/board/${id}`,
+            }).done(function () {
+                alert('삭제가 완료되었습니다');
+                window.history.back();
+            }).fail(function (error){
+                alert(JSON.stringify(error));
+            })
+        } else {
+            return false;
+        }
+     }
 };
 
-main.init();
+board.init();
